requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.8.2
MarkupSafe==3.0.2
oauthlib==3.2.2
openai==1.58.1
pydantic==2.10.4
pydantic-core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.3.1
sniffio==1.3.1
tqdm==4.67.1
tweepy==4.14.0
typing-extensions==4.12.2
urllib3==2.2.3
werkzeug==3.1.3
zipp==3.21.0
